Record the selection manifest on the portal conversion path
The CLI path wrote output/conversion-selection.json but the portal path
did not, so a focused conversion started from the modal left no record of
which programs were selected, why each was included, or which CALL targets
went unresolved. Without it a narrow run cannot be told apart from a
whole-estate run that lost programs, and cannot be reproduced.

The manifest writer moves out of ResolveProgramsCommand into the facts
infrastructure so both front doors emit the same schema from one
implementation, as staging already does.
